refactor: create Common folder for common components
