+++ wealth-sync/src/components/Common/LoadingText.tsx
