+++ wealth-sync/src/components/UserNav.tsx
@@ -2,7 +2,7 @@
 
 import { useSession, signOut } from "next-auth/react";
 import { Button } from "./ui/button";
-import LoadingText from "./LoadingText";
+import LoadingText from "./Common/LoadingText";
 
 export function UserNav() {
   const { data: session, status } = useSession();
